sistema_pedidos/requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.6
Django==6.0.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
et_xmlfile==2.0.0
inflection==0.5.1
mysqlclient==2.2.8
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
PyJWT==2.12.1
pytz==2026.1.post1
PyYAML==6.0.3
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
gunicorn==25.3.0
whitenoise==6.12.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
